Fixed sum-of-parts resistance bug in DcMotorExFake
The bug was as follows: If a motor could apply a force to produce 100
tick/sec velocity, a set velocity of 30 tick/sec would be maintained
even with a 60 tick/sec counter-rotation from addAngularVel(). If working
as expected, when another 60 tick/sec counter-rotation is applied, the
velocity should not hold, and only -30 tick/sec velocity would be
maintained. However, using addAngularVel() twice in this way would still
maintain an incorrect 30 tick/sec. This has been fixed, and now a -30
tick/sec velocity is held, as expected.