Defined Updater interface so as to update fakes.
The expected implementation of an updater is a LynxModuleFake which
registers all controllers (or devices if that is better) and updates
them upon request OR when an automatic request is done. Automatic
requests can simulate bulk caching modes, too.

It is 11:10 PM (23:10 for Commies) as I write this. I didn't have plans
tomorrow... w a i t, no I DO.