Initiated LynxUsbDeviceImplFake for LynxModules
It is intended that LynxModules are created by using the getOrAddModule
method on LynxUsbDeviceImplFake. Furthermore, LynxUsbDeviceImplFakes
require that motors (and later ananlog input and digital channels) be
added to the object for the sake of bulk reads.

LynxUsbDeviceImplFake is not an Updateable nor an Updater. Its purpose
is the construction of LynxModules and the interpretation of commands
from such.